+++ Sources/TDDKit/XCTCaptureError.swift
@@ -44,3 +44,49 @@ public extension XCTestCase {
         }
     }
 }
+
+// MARK: - synchronous
+public extension XCTestCase {
+    /// Capture the error from a throwing function call.
+    ///
+    /// This function will return the error from a throwing call.
+    /// If no error was thrown, it will record a test failure.
+    ///
+    /// ```swift
+    /// func test_failingGetUsername_createPoem_fails() throws {
+    ///     let error = XCTAnyError()
+    ///     let (sut, _) = makeSUT(getUsernameResult: .failure(error))
+    ///
+    ///     let capturedError = XCTCaptureError(from: try sut.createPoem())
+    ///
+    ///     XCTAssertCastEqual(capturedError, error)
+    /// }
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - block: The function call under test.
+    ///   - message: An optional description of a failure.
+    ///   - file: The file where the failure occurs.
+    ///   The default is the filename of the test case where you call this function.
+    ///   - line: The line number where the failure occurs.
+    ///   The default is the line number where you call this function.
+    /// - Returns: The error thrown from the block or nil if no error thrown.
+     func XCTCaptureError<T>(
+        from block: @autoclosure () throws -> T,
+        _ message: @autoclosure () -> String = "",
+        file: StaticString = #file,
+        line: UInt = #line
+     ) -> Error? {
+        do {
+            _ = try block()
+            let description = ["XCTCaptureError failed: should have thrown an error", message()]
+                .filter { !$0.isEmpty }
+                .joined(separator: " - ")
+            let context: XCTSourceCodeContext = .init(location: .init(filePath: file, lineNumber: line))
+            record(.init(type: .assertionFailure, compactDescription: description, sourceCodeContext: context))
+            return nil
+        } catch {
+            return error
+        }
+    }
+}